feat: add remove account and update password
