+++ src/assets/localisation/ru/translate.json
@@ -75,8 +75,9 @@
     "pass": "Текущий Пароль:",
     "alert": {
       "success": "Профиль успешно обновлен",
+      "successPass": "Пароль успешно обновлен",
       "error": "Не удалось обновить профиль",
-      "errorPass": "Неверноый пароль"
+      "errorPass": "Неверный пароль"
     },
     "edit": {
       "title": "Сменить Пароль",
